requirements.txt
boto3==1.12.28
botocore==1.15.28
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
docutils==0.15.2
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
pdf2image==1.12.1
Pillow==7.0.0
pytesseract==0.3.3
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
